refactor(commonjs): Extract helpers
