+++ packages/commonjs/src/parse.js
@@ -1,0 +1,17 @@
+export function tryParse(parse, code, id) {
+  try {
+    return parse(code, { allowReturnOutsideFunction: true });
+  } catch (err) {
+    err.message += ` in ${id}`;
+    throw err;
+  }
+}
+
+const firstpassGlobal = /\b(?:require|module|exports|global)\b/;
+
+const firstpassNoGlobal = /\b(?:require|module|exports)\b/;
+
+export function hasCjsKeywords(code, ignoreGlobal) {
+  const firstpass = ignoreGlobal ? firstpassNoGlobal : firstpassGlobal;
+  return firstpass.test(code);
+}
